Downloaded code. Needs modification
None of the code in this is my original work. I need to figure out how to modify it for the latch lock.  The only thing I changed in the .ino file is just adding in the youtube link of the original project.
Also, the HTML file is where the online facial recognition page is for this particular given code.